refactor: 카드 1장 추가 로직 수정
