+++ src/test/java/model/card/CardTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
